use networkpolicy
